#188 <Coupon> 쿠폰 발행한 유저의 데이터베이스에 요청한 쿠폰 데이터 저장

+++ com.taken_seat.coupon-service/src/main/java/com/taken_seat/coupon_service/presentation/dto/UpdateCouponRequestDto.java
@@ -20,7 +20,7 @@ public class UpdateCouponRequestDto {
 
     private Optional<String> code;
 
-    private Optional<Integer> quantity;
+    private Optional<Long> quantity;
 
     private Optional<Integer> discount;
